requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
dnspython==2.2.1
fastapi==0.85.1
h11==0.14.0
idna==3.4
motor==3.1.1
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
